+++ Robust.Client/UserInterface/CustomControls/DebugConsole.xaml.Completions.cs
@@ -276,10 +276,17 @@ public sealed partial class DebugConsole
                 // This means that letter casing will match the completion suggestion.
                 CommandBar.CursorPosition = lastRange.end;
                 CommandBar.SelectionStart = lastRange.start;
-                var insertValue = CommandParsing.Escape(completion);
+                var insertValue = completion;
+                var mustQuote = false;
 
-                // If the replacement contains a space, we must quote it to treat it as a single argument.
-                var mustQuote = insertValue.Contains(' ');
+                if ((completionFlags & CompletionOptionFlags.Literal) == 0)
+                {
+                    insertValue = CommandParsing.Escape(completion);
+
+                    // If the replacement contains a space, we must quote it to treat it as a single argument.
+                    mustQuote = insertValue.Contains(' ');
+                }
+
                 if ((completionFlags & CompletionOptionFlags.PartialCompletion) == 0)
                 {
                     if (mustQuote)